Update and rename OpenCVTest.java to samplePipeline.java
